test: concentrate source semantic owners
Collapse the remaining duplicated Phase 2 source/provider owners into the canonical semantic and parser-law files, move version coverage out of UI tests, and remove the semantic hitchhiker block from core filter tests.
