+++ tests/unit/core/test_filters.py
@@ -10,22 +10,6 @@ import pytest
 
 from polylogue.lib.filters import ConversationFilter
 from polylogue.lib.models import Conversation, ConversationSummary, Message
-from polylogue.lib.provider_semantics import (
-    extract_chatgpt_text,
-    extract_claude_code_text,
-    extract_codex_text,
-)
-from polylogue.schemas.unified import (
-    HarmonizedMessage,
-    bulk_harmonize,
-    extract_content_blocks,
-    extract_from_provider_meta,
-    extract_harmonized_message,
-    extract_reasoning_traces,
-    extract_token_usage,
-    harmonize_parsed_message,
-    is_message_record,
-)
 from polylogue.schemas.validator import (
     SchemaValidator,
     ValidationResult,
@@ -100,59 +84,6 @@ PICK_OPERATION_CASES = [
     (True, True, "builtins.input", EOFError, "EOFError handling"),
     (True, True, "builtins.input", KeyboardInterrupt, "KeyboardInterrupt handling"),
 ]
-
-REASONING_TRACES_CASES = [
-    (None, "claude", 0, None, "empty_none_content"),
-    (["string", 123], "claude", 0, None, "non_dict_block"),
-    ([{"type": "thinking", "thinking": "Let me think..."}], "claude", 1, "Let me think...", "thinking_block"),
-    ([{"isThought": True, "text": "Gemini thinking"}], "gemini", 1, "Gemini thinking", "gemini_thought"),
-    ([{"type": "thinking", "text": "Fallback text"}], "claude", 1, "Fallback text", "thinking_fallback"),
-]
-
-CONTENT_BLOCKS_CASES = [
-    (None, 0, [], "empty_none"),
-    (["string", 123, None], 0, [], "non_dict_items"),
-    ([{"type": "text", "text": "Hello"}], 1, ["text"], "text_block"),
-    ([{"type": "thinking", "thinking": "Thought"}], 1, ["thinking"], "thinking_block"),
-    ([{"type": "tool_use", "name": "bash", "id": "tool1", "input": {"command": "ls"}}], 1, ["tool_use"], "tool_use_block"),
-    ([{"type": "tool_result", "content": "result data"}], 1, ["tool_result"], "tool_result_block"),
-    ([{"type": "tool_result"}], 1, ["tool_result"], "tool_result_no_content"),
-    ([{"type": "code", "text": "print('hello')", "language": "python"}], 1, ["code"], "code_block_text"),
-    ([{"type": "code", "code": "def test(): pass"}], 1, ["code"], "code_block_code"),
-    ([{"type": "unknown", "data": "something"}], 0, [], "unknown_block_type"),
-]
-
-EXTRACT_TEXT_CASES = [
-    (extract_claude_code_text, None, "", "claude_code_none"),
-    (extract_claude_code_text, ["string", 123], "", "claude_code_non_dict"),
-    (extract_chatgpt_text, None, "", "chatgpt_none"),
-    (extract_chatgpt_text, {}, "", "chatgpt_no_parts"),
-    (extract_chatgpt_text, {"parts": "string"}, "string", "chatgpt_parts_as_string"),
-    (extract_chatgpt_text, {"parts": [123, "text", {"key": "val"}]}, "text", "chatgpt_non_string_parts"),
-    (extract_codex_text, {"data": "not a list"}, "", "codex_non_list"),
-]
-
-HARMONIZED_MESSAGE_PROVIDER_CASES = [
-    ("claude-code", "claude_code_msg", "Claude Code extraction"),
-    ("claude-ai", "claude_ai_msg", "Claude AI extraction"),
-    ("chatgpt", "chatgpt_msg", "ChatGPT extraction"),
-    ("gemini", "gemini_msg", "Gemini extraction"),
-    ("codex", "codex_msg", "Codex extraction"),
-]
-
-MESSAGE_RECORD_TYPE_CASES = [
-    ("claude-code", "user", True, "Claude Code user"),
-    ("claude-code", "assistant", True, "Claude Code assistant"),
-    ("claude-code", "metadata", False, "Claude Code metadata"),
-    ("chatgpt", "anything", True, "Other providers always True"),
-]
-
-TOKEN_USAGE_CASES = [
-    (None, None, "None usage"),
-    ({}, None, "Empty dict (no tokens)"),
-    ({"input_tokens": 100}, "partial", "Partial token fields"),
-]
-
 
 @pytest.fixture
 def filter_db(tmp_path):
@@ -1242,236 +1173,6 @@ class TestFiltersNegativeFTSLogic:
                    .exclude_text("error")
                    .list())
         assert not any("error" in c.display_title or any("error" in (m.text or "").lower() for m in c.messages) for c in results)
-
-
-# =============================================================================
-# UNIFIED.PY TESTS
-# =============================================================================
-
-
-class TestUnifiedMissingRole:
-    """Test _missing_role() error."""
-
-    def test_missing_role_raises_error(self):
-        """Test _missing_role() raises ValueError."""
-        from polylogue.schemas.unified import _missing_role
-        with pytest.raises(ValueError, match="Message has no role"):
-            _missing_role()
-
-
-class TestUnifiedExtractReasoningTraces:
-    """Test extract_reasoning_traces with all branches."""
-
-    @pytest.mark.parametrize("content,provider,expected_len,expected_text,description", REASONING_TRACES_CASES)
-    def test_extract_reasoning_traces(self, content, provider, expected_len, expected_text, description):
-        """Test reasoning trace extraction with various content types."""
-        result = extract_reasoning_traces(content, provider)
-        assert len(result) == expected_len
-        if expected_text is not None:
-            assert result[0].text == expected_text
-
-
-class TestUnifiedExtractContentBlocks:
-    """Test extract_content_blocks with all block types."""
-
-    @pytest.mark.parametrize("content,expected_len,expected_types,description", CONTENT_BLOCKS_CASES)
-    def test_extract_content_blocks(self, content, expected_len, expected_types, description):
-        """Test content block extraction with various block types."""
-        result = extract_content_blocks(content)
-        assert len(result) == expected_len
-        if expected_types:
-            for block, expected_type in zip(result, expected_types, strict=True):
-                assert block.type.value == expected_type
-
-
-class TestUnifiedExtractTokenUsage:
-    """Test extract_token_usage edge cases."""
-
-    @pytest.mark.parametrize("usage,expected_type,description", TOKEN_USAGE_CASES)
-    def test_extract_token_usage(self, usage, expected_type, description):
-        """Test token usage extraction with various inputs."""
-        result = extract_token_usage(usage)
-
-        if expected_type is None:
-            assert result is None
-        elif expected_type == "partial":
-            assert result.input_tokens == 100
-
-
-class TestUnifiedExtractTextHelpers:
-    """Test text extraction helpers."""
-
-    @pytest.mark.parametrize("extract_fn,content,expected,description", EXTRACT_TEXT_CASES)
-    def test_extract_text(self, extract_fn, content, expected, description):
-        """Test text extraction from various content structures."""
-        result = extract_fn(content)
-        assert result == expected or expected in result
-
-
-class TestUnifiedExtractHarmonizedMessage:
-    """Test extract_harmonized_message with all providers."""
-
-    def test_extract_harmonized_message_invalid_provider(self):
-        """Test with unknown provider."""
-        with pytest.raises(ValueError, match="Unknown provider"):
-            extract_harmonized_message("unknown_provider", {})
-
-    @pytest.mark.parametrize("provider,msg_key,description", HARMONIZED_MESSAGE_PROVIDER_CASES)
-    def test_extract_harmonized_message_by_provider(self, provider, msg_key, description):
-        """Test extract_harmonized_message for each provider."""
-        if provider == "claude-code":
-            raw = {
-                "uuid": "msg1",
-                "timestamp": "2024-01-01T00:00:00Z",
-                "message": {
-                    "role": "user",
-                    "content": [{"type": "text", "text": "Hello"}],
-                    "model": "claude"
-                }
-            }
-        elif provider == "claude-ai":
-            raw = {
-                "uuid": "msg1",
-                "sender": "user",
-                "text": "Hello",
-                "created_at": "2024-01-01T00:00:00Z"
-            }
-        elif provider == "chatgpt":
-            raw = {
-                "id": "msg1",
-                "author": {"role": "user"},
-                "content": {"parts": ["Hello"]},
-                "create_time": 1704067200
-            }
-        elif provider == "gemini":
-            raw = {
-                "role": "user",
-                "text": "Hello"
-            }
-        elif provider == "codex":
-            raw = {
-                "id": "msg1",
-                "role": "user",
-                "content": [{"text": "Hello"}],
-                "timestamp": "2024-01-01T00:00:00Z"
-            }
-
-        result = extract_harmonized_message(provider, raw)
-        assert isinstance(result, HarmonizedMessage)
-
-
-class TestUnifiedHarmonizeParsedMessage:
-    """Test harmonize_parsed_message edge cases."""
-
-    def test_harmonize_parsed_message_none_meta(self):
-        """Test with None provider_meta."""
-        result = harmonize_parsed_message("claude", None)
-        assert result is None
-
-    def test_harmonize_parsed_message_not_message_record(self):
-        """Test with non-message record."""
-        result = harmonize_parsed_message("claude-code", {"type": "metadata"})
-        assert result is None
-
-    def test_harmonize_parsed_message_valid(self):
-        """Test valid harmonization."""
-        meta = {
-            "raw": {
-                "uuid": "msg1",
-                "sender": "user",
-                "text": "Hello"
-            }
-        }
-        result = harmonize_parsed_message("claude-ai", meta)
-        assert isinstance(result, HarmonizedMessage)
-
-    def test_harmonize_parsed_message_claude_code_raw_preserves_tool_fields(self):
-        """Claude Code raw format should extract tool id/input/category correctly."""
-        raw_record = {
-            "type": "assistant",
-            "uuid": "msg-1",
-            "timestamp": "2026-01-01T00:00:00Z",
-            "message": {
-                "role": "assistant",
-                "content": [
-                    {"type": "text", "text": "Need to read file"},
-                    {"type": "thinking", "thinking": "First inspect the repo"},
-                    {
-                        "type": "tool_use",
-                        "id": "tool-1",
-                        "name": "Read",
-                        "input": {"file_path": "README.md"},
-                    },
-                ],
-            },
-        }
-        result = harmonize_parsed_message("claude-code", {"raw": raw_record})
-        assert result is not None
-        assert result.id == "msg-1"
-        assert result.role == "assistant"
-        assert result.text == "Need to read file"
-        assert len(result.tool_calls) == 1
-        assert result.tool_calls[0].id == "tool-1"
-        assert result.tool_calls[0].input == {"file_path": "README.md"}
-        assert result.tool_calls[0].category.value == "file_read"
-        assert len(result.reasoning_traces) == 1
-        assert result.reasoning_traces[0].text == "First inspect the repo"
-
-    def test_extract_from_provider_meta_claude_code_raw_extracts_text(self):
-        """Raw format extraction should produce text from text-only blocks."""
-        raw_record = {
-            "type": "assistant",
-            "uuid": "msg-2",
-            "message": {
-                "role": "assistant",
-                "content": [
-                    {"type": "text", "text": "line one"},
-                    {"type": "text", "text": "line two"},
-                    {"type": "thinking", "thinking": "some reasoning"},
-                ],
-            },
-        }
-        result = extract_from_provider_meta("claude-code", {"raw": raw_record})
-        assert result.text == "line one\nline two"
-        assert result.role == "assistant"
-        assert len(result.reasoning_traces) == 1
-
-
-class TestUnifiedBulkHarmonize:
-    """Test bulk_harmonize edge cases."""
-
-    def test_bulk_harmonize_no_provider_meta(self):
-        """Test with parsed messages without provider_meta."""
-        class MockParsedMessage:
-            pass
-
-        messages = [MockParsedMessage()]
-        result = bulk_harmonize("claude", messages)
-        assert result == []
-
-    def test_bulk_harmonize_mixed_valid_invalid(self):
-        """Test with mix of valid and invalid records."""
-        class MockParsedMessage:
-            def __init__(self, meta=None):
-                self.provider_meta = meta
-
-        messages = [
-            MockParsedMessage({"raw": {"type": "metadata"}}),
-            MockParsedMessage({"raw": {"type": "user", "uuid": "1", "message": {"role": "user", "content": [{"type": "text", "text": "Hi"}]}}}),
-        ]
-        result = bulk_harmonize("claude-code", messages)
-        assert len(result) == 1
-
-
-class TestUnifiedIsMessageRecord:
-    """Test is_message_record for Claude Code type checking."""
-
-    @pytest.mark.parametrize("provider,record_type,expected,description", MESSAGE_RECORD_TYPE_CASES)
-    def test_is_message_record(self, provider, record_type, expected, description):
-        """Test message record type checking for various providers."""
-        result = is_message_record(provider, {"type": record_type})
-        assert result == expected
-
 
 # =============================================================================
 # VALIDATOR.PY TESTS
