+++ tests/unit/sources/test_parsers.py
@@ -332,49 +332,3 @@ This is not JSON at all, should be skipped
 			len(linked_convos) > 0
 		), "Orphaned raw record should be linked to created conversation"
 
-
-# =============================================================================
-# BOUNDARY CONDITION TESTS (PARAMETRIZED)
-# =============================================================================
-
-
-@pytest.mark.parametrize("char_count,expected_substantive", [
-	(9, False),   # Below 10-char threshold
-	(10, False),  # Exactly at threshold
-	(11, True),   # Just above threshold
-	(50, True),   # Well above
-])
-def test_substantive_message_boundary(char_count, expected_substantive):
-	"""Test substantive message length boundary (>10 chars).
-
-	CRITICAL BOUNDARY TEST that was missing from original suite.
-	"""
-	from polylogue.lib.models import Message
-
-	text = "a" * char_count
-	msg = Message(id="1", role="assistant", text=text)
-
-	assert msg.is_substantive == expected_substantive, \
-		f"Wrong is_substantive for {char_count} chars"
-
-
-@pytest.mark.parametrize("fence_markers,expected_context_dump", [
-	("``` code ```", False),   # 1 fence = 2 markers (count=2), below threshold
-	("``` code ``` ``` code ```", False),   # 2 fences = 4 markers (count=4), below threshold
-	("``` code ``` ``` code ``` ``` code ```", True),   # 3 fences = 6 markers (count=6), at threshold
-	("``` a ``` ``` b ``` ``` c ``` ``` d ```", True),   # 4 fences = 8 markers (count=8), above threshold
-])
-def test_context_dump_backtick_boundary(fence_markers, expected_context_dump):
-	"""Test context dump detection boundary (6+ backtick markers = 3+ code blocks).
-
-	CRITICAL BOUNDARY TEST that was missing from original suite.
-	The logic counts occurrences of '```' (triple backticks), not individual backtick chars.
-	A complete code fence needs opening and closing, so 3 fences = 6 markers.
-	"""
-	from polylogue.lib.models import Message
-
-	msg = Message(id="1", role="user", text=fence_markers)
-
-	assert msg.is_context_dump == expected_context_dump, \
-		f"Wrong is_context_dump for fence markers: {fence_markers}"
-
